requirements.txt
annotated-types==0.5.0
anyio==3.7.1
fastapi==0.103.1
idna==3.4
pydantic==2.3.0
pydantic_core==2.6.3
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
openai-whisper==20230314